Desktop/پوشه بابا/rehab_backend-master/clean_requirements.txt
annotated-types==0.7.0
anyio==4.9.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
boltons==25.0.0
Brotli==1.1.0
build==1.2.2.post1
certifi
cffi
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama
crispy-bootstrap4==2025.6
crispy-bootstrap5==2025.6
cryptography==45.0.3
cssselect2==0.8.0
defusedxml==0.7.1
distro
dj-database-url==3.0.0
Django==5.2.2
django-allauth==65.9.0
django-axes==8.0.0
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-debug-toolbar==5.2.0
django-extensions==4.1
django-jalali==7.4.0
django-ranged-response==0.2.0
django-recaptcha==4.1.0
django-simple-captcha==0.6.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
drf-excel==2.5.3
email_validator==2.2.0
et_xmlfile==2.0.0
fastapi==0.111.1
fastapi-cli==0.0.8
fonttools==4.58.2
frozendict==2.4.6
gunicorn==23.0.0
h11==0.14.0
html5lib==1.1
idna==3.8
importlib_metadata==8.6.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
lazy-object-proxy==1.10.0
MarkupSafe==2.1.5
numpy==1.24.3
openpyxl==3.1.5
packaging==24.1
pandas==2.0.3
Pillow==11.2.1
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycparser==2.22
pydantic==2.10.0
pydantic_core==2.27.1
Pygments==2.18.0
PyJWT==2.9.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
pyphen==0.15.0
pytest==8.3.2
python-bidi==0.4.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
reportlab==4.1.0
requests==2.32.3
requests-oauthlib==2.0.0
scikit-learn==1.4.1.post1
scipy==1.10.1
six==1.16.0
soupsieve==2.6.1
sqlparse==0.5.3
starlette==0.37.2
tablib==3.6.0
threadpoolctl==3.5.0
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.2.1
uvicorn==0.29.0
webencodings==0.5.1
weasyprint==62.0
wrapt==1.16.0
xlrd==2.0.1
XlsxWriter==3.2.0
xlwt==1.3.6
zipp==3.20.0
